pyproject.toml
[tool.poetry]
name = "omlmd"
version = "0.1.2"
description = ""
authors = ["Matteo Mortari <matteo.mortari@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
oras = "^0.1.30"
pyyaml = "^6.0.1"
nbconvert = "^7.16.4"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-mock = "^3.14.0"
jq = "^1.7.0"
scikit-learn = "^1.5.0"
ipykernel = "^6.29.4"
markdown-it-py = "^3.0.0"
model-registry = "^0.2.4a1"
ruff = "^0.5.7"
mypy = "^1.11.1"
types-pyyaml = "^6.0.12.20240808"

[tool.poetry.scripts]
omlmd = "omlmd.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
markers = [
    "e2e: end-to-end testing with localhost:5001",
]

[tool.ruff]
target-version = "py39"
respect-gitignore = true

[tool.mypy]
python_version = "3.9"
strict = false
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
ignore_missing_imports = true
